Refactor controller, close socket and fix bug
Significant refactor of the controller. The dictionaries are now
indexed by the name rather than by a socket. Baddies and police are
split in different dictionaries to aid in different movement
implementations.

Fixed a bug where we had an exception after unsubscribing because the
callback was continuing to execute even after we detected that we
needed to unsubscribe

Close socket after registering